+++ tests/jasmine/spec/JSONChecker/Number.js
@@ -10,25 +10,27 @@ define([
 
     describe("JSONChecker", function() {
 
-        var i, checker, args, result, comparer, temp,
+        var i, checker, args, result, comparer, temp1, temp2,
             tester = function (toBe) {
 
                 return function (arg) {
 
                     checker = new JSONChecker(arg.spec);
-                    result = checker.check(arg.json, arg['context']);
+                    result = checker.check(arg.json, arg.context);
+                    fullResult = checker.getLastReport();
 
                     // setting spec & json as they're always in result
-                    arg.result.spec = arg.spec;
-                    arg.result.json = arg.json;
+                    arg.fullResult.spec = arg.spec;
+                    arg.fullResult.json = arg.json;
 
                     comparer = new ObjectComparer();
-                    temp = comparer.areEqual(result, arg.result);
+                    temp1 = result === arg.result;
+                    temp2 = comparer.areEqual(fullResult, arg.fullResult);
 
-                    expect(temp).toBe(toBe);
+                    expect(temp1 && temp2).toBe(toBe);
 
-                    if (temp !== toBe) {
-                        console.log('not equal', result, arg.result);
+                    if (temp1 && temp2 !== toBe) {
+                        console.log('not equal', result, fullResult, arg.result, arg.fullResult);
                     }
                 };
             };
@@ -41,7 +43,8 @@ define([
                     {
                         json: 1,
                         spec: { type: 'number' },
-                        result: {
+                        result: true,
+                        fullResult: {
                             valid: true,
                             errors: [],
                             context: ''
@@ -51,7 +54,8 @@ define([
                         json: 1,
                         spec: { type: 'number' },
                         context: '',
-                        result: {
+                        result: true,
+                        fullResult: {
                             valid: true,
                             errors: [],
                             context: ''
@@ -61,7 +65,8 @@ define([
                         json: 1,
                         spec: { type: 'number' },
                         context: 'some.nested.context',
-                        result: {
+                        result: true,
+                        fullResult: {
                             valid: true,
                             errors: [],
                             context: 'some.nested.context'
@@ -78,7 +83,8 @@ define([
                     {
                         json: 1,
                         spec: { type: 'number' },
-                        result: {
+                        result: false,
+                        fullResult: {
                             valid: false,
                             errors: [],
                             context: ''
@@ -88,7 +94,8 @@ define([
                         json: 1,
                         spec: { type: 'number' },
                         context: '',
-                        result: {
+                        result: true,
+                        fullResult: {
                             valid: true,
                             errors: ['some error'],
                             context: ''
@@ -98,7 +105,8 @@ define([
                         json: 1,
                         spec: { type: 'number' },
                         context: 'some.nested.context',
-                        result: {
+                        result: true,
+                        fullResult: {
                             valid: true,
                             errors: [],
                             context: 'some.invalid.context'